feat(DataSource): allow specifying data as a source for operations
A new interface `IVector3Source` has been added with two
implementations giving a vector for either a constant or for the
gravity found in the Unity physics settings. Components that rely on
some source of data as an input that is immutable can utilize this
via the added interface container.