feat(BodyRepresentation): add prefab for body representation
A new prefab has been added that encapsulates and sets up a body
representation. It additionally handles any necessary configuration
of an optionally provided teleporter to utilize when the body walks
up a slope or falls down to move the play area so the headset
follows the virtual body in that scenario.